Fixed Stuff
Moved InterestingFacts class to this feature, and changed the Person class to have a field of the feature
